Moved `with_default` to _utils.py
