+++ odc/stac/_stac_load.py
@@ -33,8 +33,8 @@ from odc.geo import CRS, MaybeCRS, SomeResolution
 from odc.geo.geobox import GeoBox, GeoboxAnchor, GeoboxTiles
 from odc.geo.types import Unset
 
-from ._mdtools import ConversionConfig, output_geobox, parse_items, with_default
-from ._utils import SizedIterable, pmap
+from ._mdtools import ConversionConfig, output_geobox, parse_items
+from ._utils import SizedIterable, pmap, with_default
 from .loader import (
     DaskGraphBuilder,
     LoadChunkTask,
